FFIT-15 se modificaron los archivos de visita a visita_cliente, incluyendo el nombre de las clases
